+++ back/.gitignore
@@ -2,3 +2,5 @@
 
 # dependencies
 /node_modules/
+package.json
+package-lock.json
